+++ module-1/capstone/VendoMatic/src/main/java/com/techelevator/VendingItem.java
@@ -1,13 +1,15 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public abstract class VendingItem {
 	
 	protected String name;
-	protected double price;
+	protected BigDecimal price;
 	protected int quantity;
 	protected String uniqueID;
 	
-	public VendingItem(String name, double price, int quantity, String uniqueID) {
+	public VendingItem(String name, BigDecimal price, int quantity, String uniqueID) {
 		this.name = name;
 		this.price = price;
 		this.quantity = quantity;
